requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
langchain==0.3.14
langchain-core==0.3.29
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-chroma==0.2.0
langgraph==0.2.61
chromadb==0.5.23
sentence-transformers==3.3.1
Pillow==11.0.0
requests==2.32.3
